continue

+++ Mini_Game_251210/Mini_Game_251210/main.cpp
@@ -2,23 +2,26 @@
 #include "GameManager.h"
 
 
-GameManager* GameManager::Instance = nullptr;
+shared_ptr<GameManager> GameManager::Instance = nullptr;
 
 int main()
 {
-	
-	GameManager::Initialize();
 
-	while (1)
+	GameManager::Initialize();
+	if (auto gameManager = GameManager::GetInstance().lock())
 	{
-		GameManager::GetInstance()->Input();
-		GameManager::GetInstance()->Update();
-		if (GameManager::GetInstance()->CheckWin())
-			break;
-		if (GameManager::GetInstance()->CheckLose())
-			break;
-		GameManager::GetInstance()->Render();
+		while (1)
+		{
+			gameManager->Input();
+			gameManager->Update();
+
+			if (GameManager::GetInstance()->CheckWin())
+				break;
+			if (GameManager::GetInstance()->CheckLose())
+				break;
+			GameManager::GetInstance()->Render();
+		}
+		GameManager::GetInstance()->DeleteMember();
 	}
-	GameManager::GetInstance()->DeleteMember();
 	return 0;
 }
